Json Keys with null or empty should be removed. They add no value.
The config in the application.properties does not work or the ObjectMapper which is configured here is not used in the underlying Spring Magic. I debuged it and it looked like it but it did not work. Null Values are mapped to NullNode type which I do not want. If somebody gets this working we can remove the removeNullNodes method.