[wip] Upgrade to Spring Boot 1.5.7 and Elasticsearch 2.4.6; see below for details
---
Update "steckbrief" and "changeset" field mappings:
* migrate to new Elasticsearch 2 mappings
* change "steckbrief" mapping of "links" from "nested" to "object" as no correlation with these links is needed, e.g. between a homepage URL and a build chain URL
* fix analyzer for software_dependency.version ("standard" instead of "simple") so that a search can actually return some results
* unite different field name spellings of the same word: "licences" (British English) and "licenses" (American English); use the american spelling now

---
Patch Spring Data Elasticsearch's class TransportClientFactoryBean to be able to use the Delete By Query Plugin. See class TransportClientFactoryBean for details.

---
Use only a single Elasticsearch primary shard (Lucene index) for our Elasticsearch indexes (Steckbrief, Changeset). Pivio is so small regarding the amount of data and the load it handles that it is not
needed to run it on more than one Elasticsearch instance/node. With this change things get simpler again plus we get better performance with the same accuracy as we do not need to use the Distributed
Frequency Search (DFS) anymore.

---
Wait for a connection to Elasticsearch to become available after the Pivio server has been started and tries to create the Elasticsearch indices/mappings for Document and Changeset. We need this when Elasticsearch and Pivio server is started, e.g. with Docker Compose. Elasticsearch might not yet be available when Pivio server tries to access it during startup. Although the Elasticsearch client used by Pivio server retries to connect, it throws an exception if it cannot connect to any of its known Elasticsearch nodes (we only have one). This exception resulted in a shutdown of Pivio server as it was thrown during Pivio server's startup process. Solution: During startup we wait for 5s, then 10s, and finally 20s on a connection to Elasticsearch. While waiting the Elasticsearch client retries to connect every 5s. If we cannot get a connection within 35s Pivio server shuts down. See class ElasticsearchConnectionAvailableChecker for details.

---
Use the official Elasticsearch 2.4.6 Docker image instead of building it ourselves. That is, we do not need to maintain the code anymore, save build time, can run tests easier, and prepare for elastic's way of handle Docker Elasticsearch 5 images.

---
Replace deprecated and big-sized Debian based Oracle JRE 8 Docker base image with the up-to-date and small-sized Alpine Linux based OpenJDK 8 JRE Docker base image.

---
Use Elasticsearch's TransportClient also for tests (instead of the NodeClient). This means we have to do the lifecycle management of Elasticsearch ourselves: start it before running tests and stop it afterwards. Docker with TestContainers comes in handy. See class DockerEnvironmentInitializer for details.

---
Change integration tests (*ApiTest) so that it really tests what is going into production: the pivio-server Docker container and the Elasticsearch 2.4.6 Docker container with the installed Delete By Query Plugin. Clean up ApiTests.

---
Update Gradle to version 4.1, clean up build.gradle, and bump up dependency versions.

---
Remove @Autowired annotations on beans that only define one constructor as they have become obsolete with Spring 4.3.

---
Update documentation and URLs pointing to the Elasticsearch's guide; fix broken cURL statements.